feat(paper-trades): split paper trades into 3 components
